Refactor: Moved the WorldManager from libtiled to libtilededitor
At least as it stands, this class is only relevant in the editor. The
world load and save functions, previously part of WorldManager, are now
part of World.
