+++ src/tiled/mapeditor.cpp
@@ -50,7 +50,6 @@
 #include "objectreferencetool.h"
 #include "objectsdock.h"
 #include "objectselectiontool.h"
-#include "objecttemplate.h"
 #include "preferences.h"
 #include "propertiesdock.h"
 #include "reversingproxymodel.h"
@@ -73,10 +72,9 @@
 #include "undodock.h"
 #include "wangbrush.h"
 #include "wangdock.h"
-#include "wangset.h"
 #include "zoomable.h"
-#include "worldmovemaptool.h"
 #include "worldmanager.h"
+#include "worldmovemaptool.h"
 
 #include <QComboBox>
 #include <QDialogButtonBox>
